pgcryptocipherccl: move pgcryptocipher package to CCL
Release note: None
