+++ .github/CODEOWNERS
@@ -373,6 +373,7 @@
 /pkg/ccl/multitenant/tenantcostserver/ @cockroachdb/sqlproxy-prs
 /pkg/ccl/oidcccl/            @cockroachdb/obs-inf-prs
 /pkg/ccl/partitionccl/       @cockroachdb/sql-foundations
+/pkg/ccl/pgcryptoccl/        @cockroachdb/sql-foundations
 
 #!/pkg/ccl/serverccl/        @cockroachdb/unowned
 /pkg/ccl/serverccl/diagnosticsccl/ @cockroachdb/obs-inf-prs
