+++ pkg/BUILD.bazel
@@ -66,6 +66,7 @@ ALL_TESTS = [
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
     "//pkg/ccl/oidcccl:oidcccl_test",
     "//pkg/ccl/partitionccl:partitionccl_test",
+    "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
     "//pkg/ccl/serverccl/adminccl:adminccl_test",
     "//pkg/ccl/serverccl/diagnosticsccl:diagnosticsccl_test",
@@ -538,7 +539,6 @@ ALL_TESTS = [
     "//pkg/sql/schemachanger/screl:screl_test",
     "//pkg/sql/schemachanger/scrun:scrun_test",
     "//pkg/sql/schemachanger:schemachanger_test",
-    "//pkg/sql/sem/builtins/pgcrypto/pgcryptocipher:pgcryptocipher_test",
     "//pkg/sql/sem/builtins/pgformat:pgformat_test",
     "//pkg/sql/sem/builtins:builtins_disallowed_imports_test",
     "//pkg/sql/sem/builtins:builtins_test",
@@ -869,6 +869,8 @@ GO_TARGETS = [
     "//pkg/ccl/oidcccl:oidcccl_test",
     "//pkg/ccl/partitionccl:partitionccl",
     "//pkg/ccl/partitionccl:partitionccl_test",
+    "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl",
+    "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
     "//pkg/ccl/serverccl/adminccl:adminccl_test",
@@ -2039,8 +2041,6 @@ GO_TARGETS = [
     "//pkg/sql/sem/asof:asof",
     "//pkg/sql/sem/builtins/builtinconstants:builtinconstants",
     "//pkg/sql/sem/builtins/builtinsregistry:builtinsregistry",
-    "//pkg/sql/sem/builtins/pgcrypto/pgcryptocipher:pgcryptocipher",
-    "//pkg/sql/sem/builtins/pgcrypto/pgcryptocipher:pgcryptocipher_test",
     "//pkg/sql/sem/builtins/pgformat:pgformat",
     "//pkg/sql/sem/builtins/pgformat:pgformat_test",
     "//pkg/sql/sem/builtins:builtins",
