+++ pkg/ccl/pgcryptoccl/pgcryptocipherccl/BUILD.bazel
@@ -1,13 +1,13 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "pgcryptocipher",
+    name = "pgcryptocipherccl",
     srcs = [
         "cipher_method.go",
         "doc.go",
         "padding.go",
     ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/sql/sem/builtins/pgcrypto/pgcryptocipher",
+    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/pgcryptoccl/pgcryptocipherccl",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/sql/pgwire/pgcode",
@@ -18,13 +18,14 @@ go_library(
 )
 
 go_test(
-    name = "pgcryptocipher_test",
+    name = "pgcryptocipherccl_test",
     srcs = [
         "cipher_method_test.go",
         "padding_test.go",
     ],
     args = ["-test.timeout=295s"],
-    embed = [":pgcryptocipher"],
+    embed = [":pgcryptocipherccl"],
+    tags = ["ccl_test"],
     deps = [
         "//pkg/util/leaktest",
         "@com_github_stretchr_testify//require",
