+++ pkg/ccl/pgcryptoccl/pgcryptocipherccl/doc.go
@@ -1,0 +1,11 @@
+// Copyright 2023 The Cockroach Authors.
+//
+// Licensed as a CockroachDB Enterprise file under the Cockroach Community
+// License (the "License"); you may not use this file except in compliance with
+// the License. You may obtain a copy of the License at
+//
+//     https://github.com/cockroachdb/cockroach/blob/master/licenses/CCL.txt
+
+// Package pgcryptocipherccl contains the implementation of pgcrypto
+// cipher functions.
+package pgcryptocipherccl
