--- pkg/sql/sem/builtins/pgcrypto/pgcryptocipher/doc.go
@@ -1,13 +1,0 @@
-// Copyright 2023 The Cockroach Authors.
-//
-// Use of this software is governed by the Business Source License
-// included in the file licenses/BSL.txt.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0, included in the file
-// licenses/APL.txt.
-
-// Package pgcryptocipher contains the implementation of pgcrypto
-// cipher functions.
-package pgcryptocipher
